+++ src/components/Layout/ResultLayout.vue
@@ -48,7 +48,6 @@ export default {
   padding-top: 1em;
   min-width: 40em;
   max-width: 94%;
-  flex: 1 0 auto;
   margin-left: 1em;
 }